refactor: move getSchemaElement to own file
